backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
spotipy>=2.23.0
yt-dlp>=2024.8.0
requests>=2.31.0
gunicorn>=21.2.0
